+++ oshi-core/src/main/java/oshi/software/common/AbstractOperatingSystem.java
@@ -27,8 +27,12 @@ import static oshi.software.os.OperatingSystem.ProcessFiltering.ALL_PROCESSES;
 import static oshi.software.os.OperatingSystem.ProcessSorting.NO_SORTING;
 import static oshi.util.Memoizer.memoize;
 
+import java.util.Collection;
 import java.util.Comparator;
 import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
+import java.util.Set;
 import java.util.function.Predicate;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
@@ -113,6 +117,79 @@ public abstract class AbstractOperatingSystem implements OperatingSystem {
     }
 
     protected abstract List<OSProcess> queryChildProcesses(int parentPid);
+
+    @Override
+    public List<OSProcess> getDescendantProcesses(int parentPid, Predicate<OSProcess> filter,
+            Comparator<OSProcess> sort, int limit) {
+        return queryDescendantProcesses(parentPid).stream().filter(filter == null ? ALL_PROCESSES : filter)
+                .sorted(sort == null ? NO_SORTING : sort).limit(limit > 0 ? limit : Long.MAX_VALUE)
+                .collect(Collectors.toList());
+    }
+
+    protected abstract List<OSProcess> queryDescendantProcesses(int parentPid);
+
+    /**
+     * Utility method for subclasses to take a full process list as input and return
+     * the children or descendants of a particular process.
+     *
+     * @param allProcs
+     *            A collection of all processes
+     * @param parentPid
+     *            The process ID whose children or descendants to return
+     * @param descendantPids
+     *            On input, an empty set, or a set of other children already
+     *            retrieved when using recursion. On output, the children of the
+     *            parent added to the set.
+     * @param recurse
+     *            If false, only gets immediate children of this process. If true,
+     *            gets all descendants.
+     */
+    protected static void addChildrenToDescendantSet(Collection<OSProcess> allProcs, int parentPid,
+            Set<Integer> descendantPids, boolean recurse) {
+        // Collect this process's children
+        Set<Integer> childPids = allProcs.stream().filter(p -> p.getParentProcessID() == parentPid)
+                .filter(p -> p.getProcessID() != parentPid).map(OSProcess::getProcessID).collect(Collectors.toSet());
+        // Add to descendant set
+        descendantPids.addAll(childPids);
+        // Recurse
+        if (recurse) {
+            for (int pid : childPids) {
+                addChildrenToDescendantSet(allProcs, pid, descendantPids, true);
+            }
+        }
+    }
+
+    /**
+     * Utility method for subclasses to take a map of pid to parent as input and
+     * return the children or descendants of a particular process.
+     *
+     * @param parentPidMap
+     *            a map of all processes with processID as key and parentProcessID
+     *            as value
+     * @param parentPid
+     *            The process ID whose children or descendants to return
+     * @param descendantPids
+     *            On input, an empty set, or a set of other children already
+     *            retrieved when using recursion. On output, the children of the
+     *            parent added to the set.
+     * @param recurse
+     *            If false, only gets immediate children of this process. If true,
+     *            gets all descendants.
+     */
+    protected static void addChildrenToDescendantSet(Map<Integer, Integer> parentPidMap, int parentPid,
+            Set<Integer> descendantPids, boolean recurse) {
+        // Collect this process's children
+        Set<Integer> childPids = parentPidMap.entrySet().stream().filter(e -> e.getValue().equals(parentPid))
+                .map(Entry::getKey).filter(k -> !k.equals(parentPid)).collect(Collectors.toSet());
+        // Add to descendant set
+        descendantPids.addAll(childPids);
+        // Recurse
+        if (recurse) {
+            for (int pid : childPids) {
+                addChildrenToDescendantSet(parentPidMap, pid, descendantPids, true);
+            }
+        }
+    }
 
     @Override
     public String toString() {