+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ##### New Features
 * [#1541](https://github.com/oshi/oshi/pull/1541): Expose the alias of a network interface (Windows and Linux) - [@dornand](https://github.com/dornand).
 * [#1546](https://github.com/oshi/oshi/pull/1546): Expose network interface operational status (Windows and Linux) - [@dornand](https://github.com/dornand).
+* [#1548](https://github.com/oshi/oshi/pull/1548): Add getter for descendants of a process - [@dbwiddis](https://github.com/dbwiddis).
 
 # 5.5.0 (2021-02-08), 5.5.1 (2021-02-21)